jffavaron - 13/06/13 recompilation IDL et autres

+++ src/AssistanceTouristique/Statistique.java
@@ -15,7 +15,7 @@ public final class Statistique implements org.omg.CORBA.portable.IDLEntity
     /**
      * Struct member valeurStat
      */
-    public int valeurStat;
+    public float valeurStat;
 
     /**
      * Struct member uniteStat
@@ -34,7 +34,7 @@ public final class Statistique implements org.omg.CORBA.portable.IDLEntity
      * @param valeurStat valeurStat struct member
      * @param uniteStat uniteStat struct member
      */
-    public Statistique(String libelleStat, int valeurStat, String uniteStat)
+    public Statistique(String libelleStat, float valeurStat, String uniteStat)
     {
         this.libelleStat = libelleStat;
         this.valeurStat = valeurStat;
